# openapi_patient_service.yaml
openapi: 3.0.0
info:
  title: Patient Service API
  version: 1.0.0
  description: |
    Hospital Management System - Patient Service API.

    Responsibilities:
    - Patient CRUD
    - Search by name or phone
    - Pagination and filtering
    - Standard error response structure
    - Role-based access control
    - Correlation ID propagation

servers:
  - url: http://localhost:9001
    description: Local development server

tags:
  - name: Health
    description: Health and readiness endpoints
  - name: Patients
    description: Patient management APIs

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: MockToken

  parameters:
    CorrelationIdHeader:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        type: string
      description: Optional correlation ID for tracing requests across services

  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "404"
        message:
          type: string
          example: "Patient not found"
        correlationId:
          type: string
          example: "0f9b7c85-a530-4eb9-a3d0-5414a9364a1e"

    ValidationErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: "422"
        message:
          type: string
          example: "Validation Error"
        details:
          type: array
          items:
            type: object
        correlationId:
          type: string
          example: "0f9b7c85-a530-4eb9-a3d0-5414a9364a1e"

    Patient:
      type: object
      properties:
        patient_id:
          type: integer
          example: 1
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
        phone:
          type: string
          example: "+91-9876543210"
        dob:
          type: string
          format: date
          example: "1990-01-15"
        is_active:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: "2026-04-19T10:00:00Z"

    PatientCreate:
      type: object
      required:
        - name
        - email
        - phone
        - dob
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
        phone:
          type: string
          minLength: 10
          maxLength: 20
          example: "+91-9876543210"
        dob:
          type: string
          format: date
          example: "1990-01-15"

    PatientUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 100
          example: John Updated
        email:
          type: string
          format: email
          example: john.updated@example.com
        phone:
          type: string
          minLength: 10
          maxLength: 20
          example: "+91-9876500000"
        dob:
          type: string
          format: date
          example: "1991-02-10"
        is_active:
          type: boolean
          example: true

    PatientListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Patient'
        total:
          type: integer
          example: 60
        skip:
          type: integer
          example: 0
        limit:
          type: integer
          example: 10

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
      responses:
        '200':
          description: Service health status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  service:
                    type: string
                    example: patient

  /ready:
    get:
      tags:
        - Health
      summary: Readiness check
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
      responses:
        '200':
          description: Service readiness status
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  service:
                    type: string
                    example: patient
                  ready:
                    type: boolean
                    example: true

  /v1/patients:
    get:
      tags:
        - Patients
      summary: List patients
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - name: name
          in: query
          required: false
          schema:
            type: string
          description: Filter by patient name
        - name: phone
          in: query
          required: false
          schema:
            type: string
          description: Filter by phone
        - name: is_active
          in: query
          required: false
          schema:
            type: boolean
          description: Filter by active status
        - name: skip
          in: query
          required: false
          schema:
            type: integer
            default: 0
            minimum: 0
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: Paginated patient list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PatientListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'
        '500':
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - Patients
      summary: Create a patient
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientCreate'
      responses:
        '201':
          description: Patient created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate email conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'

  /v1/patients/{patient_id}:
    get:
      tags:
        - Patients
      summary: Get patient by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - name: patient_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Patient found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags:
        - Patients
      summary: Update patient by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - name: patient_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientUpdate'
      responses:
        '200':
          description: Patient updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate email conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationErrorResponse'

    delete:
      tags:
        - Patients
      summary: Deactivate patient by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/CorrelationIdHeader'
        - name: patient_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Patient deactivated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: Successfully deactivated
                  patient_id:
                    type: integer
                    example: 1
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Patient not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'